requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
pymongo==4.6.1
